chore: move vite config to configs dir
