+++ configs/vite.config.ts
